+++ modules/generator/registry/registry.go
@@ -125,13 +125,21 @@ func New(cfg *Config, overrides Overrides, tenant string, appendable storage.App
 	return r
 }
 
+func (r *ManagedRegistry) NewLabelValues(values []string) *LabelValues {
+	return newLabelValuesWithMax(values, r.cfg.MaxLabelValueLength)
+}
+
 func (r *ManagedRegistry) NewCounter(name string, labels []string) Counter {
+	truncateLength(labels, r.cfg.MaxLabelNameLength)
+
 	c := newCounter(name, labels, r.onAddMetricSeries, r.onRemoveMetricSeries)
 	r.registerMetric(c)
 	return c
 }
 
 func (r *ManagedRegistry) NewHistogram(name string, labels []string, buckets []float64) Histogram {
+	truncateLength(labels, r.cfg.MaxLabelNameLength)
+
 	h := newHistogram(name, labels, buckets, r.onAddMetricSeries, r.onRemoveMetricSeries)
 	r.registerMetric(h)
 	return h