+++ tempodb/encoding/vparquet/block_findtracebyid.go
@@ -93,8 +93,10 @@ func findTraceByID(ctx context.Context, traceID common.ID, meta *backend.BlockMe
 
 	// Cache of row group bounds
 	rowGroupMins := make([]common.ID, numRowGroups+1)
-	rowGroupMins[0] = meta.MinID
-	rowGroupMins[numRowGroups] = meta.MaxID // This is actually inclusive and the logic is special for the last row group below
+	// todo: restore using meta min/max id once it works
+	//    https://github.com/grafana/tempo/issues/1903
+	rowGroupMins[0] = bytes.Repeat([]byte{0}, 16)
+	rowGroupMins[numRowGroups] = bytes.Repeat([]byte{255}, 16) // This is actually inclusive and the logic is special for the last row group below
 
 	// Gets the minimum trace ID within the row group. Since the column is sorted
 	// ascending we just read the first value from the first page.